js/xpconnect/tests/chrome/test_bug763343.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=763343
-->
<window title="Mozilla Bug 763343"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=763343"
     target="_blank">Mozilla Bug 763343</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript">
  <![CDATA[

  /** Test for Cross-compartment nsIClassInfo singleton wrapping. **/
  const Cc = Components.classes;
  const Ci = Components.interfaces;
  const Cu = Components.utils;

  // We need an object here that has singleton classinfo.  For now, the console
  // service works.
  var singleton = Cc["@mozilla.org/consoleservice;1"].getService(Ci.nsIClassInfo);
  ok(singleton.flags & Ci.nsIClassInfo.SINGLETON_CLASSINFO,
     "Should have singleton classinfo");
  var sb = new Cu.Sandbox(window);

  // Don't crash.
  sb.singleton = singleton;
  ok(true, "didn't crash");

  ]]>
  </script>
</window>
